Revert "Make ddev tests more deterministic (#16994)"
This reverts commit a4b4ec0b4a887fe1da3a19b8684e37602336c3d8.
